Stub instructions format
