+++ src/cpu/mos6502/operations/mnemonic.rs
@@ -77,30 +77,16 @@ pub struct CLV;
 pub struct BRK;
 pub struct NOP;
 
-impl<'a> Parser<'a, &'a [u8], NOP> for NOP {
-    fn parse(&self, _: &'a [u8]) -> ParseResult<&'a [u8], NOP> {
-        todo!()
+impl From<NOP> for u8 {
+    fn from(_: NOP) -> Self {
+        0xea
     }
 }
 
-/// Address modes
-pub struct Accumulator;
-pub struct Implied;
-pub struct Immediate(u8);
-pub struct Absolute(u16);
-pub struct ZeroPage(u8);
-pub struct Relative(i8);
-pub struct Indirect(u16);
-pub struct AbsoluteIndexedWithX(u16);
-pub struct AbsoluteIndexedWithY(u16);
-pub struct ZeroPageIndexedWithX(u8);
-pub struct ZeroPageIndexedWithY(u8);
-pub struct IndexedIndirect(u8);
-pub struct IndirectIndexed(u8);
-
-#[derive(Copy, Clone, PartialEq, Debug)]
-/// Operation takes a mnemonic
-pub struct Operation<M, A> {
-    mnemonic: M,
-    address_mode: A,
+impl<'a> Parser<'a, &'a [u8], NOP> for NOP {
+    fn parse(&self, input: &'a [u8]) -> ParseResult<&'a [u8], NOP> {
+        parcel::parsers::byte::expect_byte(NOP.into())
+            .map(|_| NOP)
+            .parse(input)
+    }
 }
