+++ src/cpu/mos6502/operations/mod.rs
@@ -1,0 +1,9 @@
+pub mod address_mode;
+pub mod mnemonic;
+
+#[derive(Copy, Clone, PartialEq, Debug)]
+/// Operation takes a mnemonic
+pub struct Operation<M, A> {
+    mnemonic: M,
+    address_mode: A,
+}
